chore: update c7n to 0.9.44, prep release 0.5.30 (ENG-5015) (#56)
* chore: update python version matrix and setup-python action

* chore: prep release 0.5.30

- Update to c7n-left 0.3.30
- Add dev dependencies for tomlkit and semver

* chore: use tomlkit in scripts/upgrade.py

tomllib is installed in Python 3.11+, but since we're
reading/writing/updating existing TOML files we need another package.
Python recommends TOML Kit for this sort of use case, and it works with
all Python versions we support.

See also: https://docs.python.org/3/library/tomllib.html

* add 3.11 back to test matrix

CI checks are stalled without it, and maybe it's better to leave anyway?

Also move from black/flake8 to ruff, to sync with some other repos and
avoid pre-commit failures on 3.13.

* nevermind, update flake8 instead

Moving to ruff makes formatting changes on 4 files. Not a huge deal, but
probably not a necessary switch for _this_ PR.